+++ WSW/app/src/main/java/com/example/wsw/MainActivity.kt
@@ -1,16 +1,15 @@
 package com.example.wsw
 
-import android.content.Intent
-import android.content.pm.PackageManager
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.view.MenuItem
-import android.widget.Toast
-import androidx.core.app.ActivityCompat
-import androidx.core.content.ContextCompat
 import androidx.fragment.app.Fragment
+import com.example.wsw.ui.home.HomeFragment
+import com.example.wsw.ui.questionboard.QuestionBoardFragment
+import com.example.wsw.ui.reviewboard.ReviewBoardFragment
+import com.example.wsw.ui.secretboard.SecretBoardFragment
+import com.example.wsw.ui.studyboard.StudyBoardFragment
 import com.google.android.material.bottomnavigation.BottomNavigationView
-import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity(), BottomNavigationView.OnNavigationItemSelectedListener {
