+++ WSW/app/src/main/java/com/example/wsw/SplashActivity.kt
@@ -5,6 +5,7 @@ import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.os.Handler
 import android.os.Looper
+import com.example.wsw.feature.login.LoginActivity
 
 
 class SplashActivity : AppCompatActivity() {
