+++ WSW/app/src/main/java/com/example/wsw/feature/signup/SignupActivity.kt
@@ -4,7 +4,7 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import android.widget.TextView
-import com.example.wsw.LoginActivity
+import com.example.wsw.feature.login.LoginActivity
 import com.example.wsw.R
 
 class SignupActivity : AppCompatActivity() {
