+++ WSW/app/src/main/java/com/example/wsw/ui/home/HomeFragment.kt
@@ -1,26 +1,22 @@
-package com.example.wsw
+package com.example.wsw.ui.home
 
 import android.content.Context
-import android.content.Intent
 import android.content.SharedPreferences
 import android.os.Bundle
 import android.util.Log
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
-import android.view.MotionEvent
 import android.view.View
 import android.view.ViewGroup
-import android.widget.ImageView
 import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.Navigation
-import kotlinx.android.synthetic.main.fragment_home.*
+import com.example.wsw.R
 
 class HomeFragment :Fragment(){
 
     companion object{
         const val TAG: String = "HomeFragment"
 
-        fun newInstance(): HomeFragment{
+        fun newInstance(): HomeFragment {
             return HomeFragment()
         }
     }
