+++ WSW/app/src/main/java/com/example/wsw/ui/questionboard/QuestionBoardFragment.kt
@@ -1,18 +1,17 @@
-package com.example.wsw
+package com.example.wsw.ui.questionboard
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import androidx.recyclerview.widget.RecyclerView
+import com.example.wsw.R
 
-class SecretBoardFragment: Fragment() {
+class QuestionBoardFragment : Fragment() {
 
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view = inflater.inflate(R.layout.fragment_secret_board, container, false)
-        
+        val view = inflater.inflate(R.layout.fragment_question_board, container, false)
+
         return view
     }
-
 }
