+++ WSW/app/src/main/java/com/example/wsw/ui/questionboard/QuestionBoardPostActivity.kt
@@ -1,13 +1,14 @@
-package com.example.wsw
+package com.example.wsw.ui.questionboard
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.wsw.R
 
-class ReviewBoardPostActivity : AppCompatActivity(){
+class QuestionBoardPostActivity : AppCompatActivity(){
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.fragment_review_board_post)
+        setContentView(R.layout.fragment_question_board_post)
 
     }
 }
