+++ WSW/app/src/main/java/com/example/wsw/ui/reviewboard/ReviewBoardFragment.kt
@@ -1,16 +1,15 @@
-package com.example.wsw
+package com.example.wsw.ui.reviewboard
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.example.wsw.R
 
-class QuestionBoardFragment : Fragment() {
-
+class ReviewBoardFragment : Fragment() {
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        val view = inflater.inflate(R.layout.fragment_question_board, container, false)
-
+        var view = LayoutInflater.from(activity).inflate(R.layout.fragment_review_board, container, false)
         return view
     }
 }
