+++ WSW/app/src/main/java/com/example/wsw/ui/reviewboard/ReviewBoardPostActivity.kt
@@ -1,13 +1,14 @@
-package com.example.wsw
+package com.example.wsw.ui.reviewboard
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.wsw.R
 
-class SecretBoardWriteActivity : AppCompatActivity(){
+class ReviewBoardPostActivity : AppCompatActivity(){
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.fragment_secret_board_write)
+        setContentView(R.layout.fragment_review_board_post)
 
     }
 }
