+++ WSW/app/src/main/java/com/example/wsw/ui/secretboard/SecretBoardFragment.kt
@@ -1,14 +1,18 @@
-package com.example.wsw
+package com.example.wsw.ui.secretboard
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
+import com.example.wsw.R
 
-class ReviewBoardFragment : Fragment() {
+class SecretBoardFragment: Fragment() {
+
     override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
-        var view = LayoutInflater.from(activity).inflate(R.layout.fragment_review_board, container, false)
+        val view = inflater.inflate(R.layout.fragment_secret_board, container, false)
+        
         return view
     }
+
 }
