+++ WSW/app/src/main/java/com/example/wsw/ui/secretboard/post/SecretBoardPostActivity.kt
@@ -1,13 +1,13 @@
-package com.example.wsw
+package com.example.wsw.ui.secretboard.post
 
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
+import com.example.wsw.R
 
-class StudyBoardPostActivity : AppCompatActivity(){
+class SecretBoardPostActivity: AppCompatActivity(){
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.fragment_study_board_post)
-
+        setContentView(R.layout.fragment_secret_board_post)
     }
 }
