+++ WSW/app/src/main/res/layout/activity_login.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".LoginActivity">
+    tools:context=".feature.login.LoginActivity">
 
     <ImageView
         android:layout_width="match_parent"
